[LayoutNG] Fix order OOFs are added in
I was investigating whether we were handling break tokens correctly
for new columns added during OOF positioned layout. This ended up
looking correct as-is, but I ran into a couple of issues while
testing this.

1. We were getting the incorrect previous_break_token for columns
in the presence of spanners. To fix this, look for the first
non-spanner sibling for the previous break token.

2. We were adding OOFs to columns in an incorrect order. Instead of
simply looking at IsFirstForNode(), we also have to look at the
fragment index of existing children vs that of the OOF
fragments being added to determine a consistent child order.

To efficiently handle to ordering problem, I've also slightly
updated how we are storing the OOF layout results to not only group them
by fragmentainer index, but also their by fragment indexes.

Bug: 1079031
Change-Id: Iaa7d5b19bc3ca101554f24a672989c6531abfce1