Initial implementation of scrollbar-width
ScrollbarTheme::ScrollbarThickness() takes scrollbar-width into
account when calculating the thickness.

Scrollbar themes return a thickness of 0 if scrollbar-width is "none"
and the same value for both "auto" and "thin", with the exception
of ScrollbarThemeMac which uses NSSmallControlSize and
NSRegularControlSize to size and display the appropriate scrollbar.

ComputedStyle::HasCustomScrollbarStyle() returns true if a custom
style for the scrollbar has been set via ::-webkit-scrollbar and
the value of scrollbar-width has not been changed from the default.
This method replaces HasPseudoElementStyle(kPseudoIdScrollbar)
throughout the codebase so the standard properties for styling
scrollbars take precedence over the non-standard method.

WPT tests are included.

Bug: 891944
Change-Id: I9f77d46fc1b05a8d05bb4eeffeef4f8e52f8031b