Portals: focus() inside portal should change active element
Allows a portal document's activeElement to update when focus is called.
Note that focus events won't be dispatched as the portal doesn't get
page focus, and the portal WebContents isn't set as the focused
WebContents (with the exception of orphaned portals). Updating the
activeElement also means an autofocused element inside a portal will
receive focus when the portal is activated.

More discussion here: https://github.com/WICG/portals/issues/257

Change-Id: If67be1f424114653983b473a80c515337d49596a
Bug: 1059404